feat: 당첨 번호 입력 및 보너스볼 입력 구현
- 출력 정리

+++ src/main/kotlin/view/InputView.kt
@@ -41,10 +41,10 @@ class InputView(private val scanner: Scanner) {
         return (START..count).map { inputLottoNumbers() }
     }
 
-    fun inputManualCount() = inputNumber().toInt()
+    fun inputManualCount() = inputNumber()
 
     fun inputNumber() = try {
-        scanner.nextLine().toLong()
+        scanner.nextLine().toInt()
     } catch (e: NumberFormatException) {
         throw NumberFormatException(NUMBER_FORMAT_MESSAGE)
     }
